pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "admin-pdf-toolkit"
version = "1.11.0"
description = "Open-source, offline PDF toolkit by Engin — convert (Excel/Word/JPG/OCR), edit (annotation/overlay/replace), 35+ tools, web UI + PWA + Windows service."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Orhan Engin Okay", email = "orhanenginokay@gmail.com" },
]
keywords = [
    "pdf",
    "toolkit",
    "editor",
    "converter",
    "ocr",
    "excel",
    "word",
    "fastapi",
    "pwa",
    "offline",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Natural Language :: English",
    "Natural Language :: Turkish",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
dependencies = [
    "pymupdf>=1.27,<2.0",
    "pdfplumber>=0.11,<1.0",
    "pdf2docx>=0.5.12,<1.0",
    "openpyxl>=3.1,<4.0",
    "python-docx>=1.1,<2.0",
    "xhtml2pdf>=0.2.16,<1.0",
    "Pillow>=10.0,<12.0",
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "python-multipart>=0.0.20,<1.0",
    "jinja2>=3.1,<4.0",
    "easyocr>=1.7,<2.0",
    "cryptography>=43.0,<46.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.urls]
Homepage = "https://github.com/orhanenginokay/pdfconverter"
Repository = "https://github.com/orhanenginokay/pdfconverter"
Issues = "https://github.com/orhanenginokay/pdfconverter/issues"
Changelog = "https://github.com/orhanenginokay/pdfconverter/blob/main/CHANGELOG.md"

[project.scripts]
ht-pdf = "app:main"

[tool.setuptools]
py-modules = ["app", "pdf_converter", "pdf_safety"]

[tool.setuptools.package-data]
"*" = ["templates/*.html", "static/*"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
    "_work",
    "output",
    "dist",
    "build",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long (formatter handles)
    "B008",  # function call in default argument (FastAPI Depends pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# ---------------------------------------------------------------------------
# mypy — type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
exclude = [
    "_work",
    "output",
    "dist",
    "build",
]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
branch = true
source = ["app", "pdf_converter", "pdf_safety"]
omit = [
    "tests/*",
    "build_portable.py",
    "_work/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
